Add clang-tidy to pre-commit hooks and CI (#26)
* Add clang-tidy to pre-commit hooks and CI

- Add .clang-tidy configuration with enabled checks for bugprone,
  clang-analyzer, misc, modernize, performance, and readability
- Add clang-tidy pre-commit hook with fallback to brew's llvm
- Add separate lint-cpp CI job that runs in parallel with build-and-test
- Update CMakeLists.txt to export compile_commands.json and set sysroot
- Fix code issues found by clang-tidy:
  - Fix narrowing conversions with explicit static_cast
  - Remove unused variables (two, neg_one, edgePaddingHigh, sliceSizes)
  - Fix nullptr usage
  - Fix container size checks (.size() >= 1 → !.empty())
  - Simplify duplicate branch logic
- Exclude pjrt_executable.cc from clang-tidy (requires generated proto)
- Add .cache/ to .gitignore


* ci: separate brew install steps to measure timing


* ci: cache llvm installation


* ci: cache llvm and clang-format together

Cache the Cellar directories and use brew link to restore symlinks
on cache hit.


* ci: fix llvm cache by recreating opt symlinks

The /opt/homebrew/opt/llvm symlink doesn't exist after restoring
the Cellar cache. Manually create the symlinks instead of using
brew link.


* ci: add clang-format to PATH for cache restore


* ci: use tecolicom/actions-use-homebrew-tools for caching

This action properly handles caching Homebrew packages including
their transitive dependencies (e.g., z3 for llvm).


* ci: skip build-deps job when cache exists

Add a lightweight check-cache job that uses lookup-only to check
if the MLIR/StableHLO cache exists. The build-deps job now only
runs on cache miss, saving ~30s on cache hit.


---------

+++ src/pjrt_plugin/mps_executable.mm
@@ -33,7 +33,7 @@ MpsExecutable::MpsExecutable(MpsClient* client, mps::ParsedModule module)
     name_ = entry_func_.getName().str();
 
     // Set the number of outputs based on result types
-    num_outputs_ = entry_func_.getNumResults();
+    num_outputs_ = static_cast<int>(entry_func_.getNumResults());
 
     valid_ = true;
 }
